requirements.txt
bcrypt==4.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
